Conexion con la base de datos

+++ src/main/java/co/edu/uniminuto/inventario/ComidaGato.java
@@ -67,9 +67,8 @@ public class ComidaGato extends Productos{
     }
 
     public void setCantidad(int cantidad) {
-        this.cantidad = cantidad;
+        this.cantidad = (short) cantidad;
     }
-
